+++ src/content/fragments/changelog.md
@@ -79,7 +79,7 @@
 - [Microlink API](/docs/api/getting-started/overview): Added [JSON](/docs/guides/data-extraction) data extraction support.
 - [Microlink API](/docs/api/getting-started/overview): Improved [markdown](/docs/api/parameters/markdown) table and code block rendering.
 - [Microlink API](/docs/api/getting-started/overview): Improved Reddit, Instagram, and LinkedIn URLs resolution.
-- [Microlink CLI](https://github.com/microlinkhq/cli): Added [headers](/docs/api/parameters/headers) support.
+- [Microlink CLI](https://github.com/microlinkhq/microlink): Added [headers](/docs/api/parameters/headers) support.
 - [unavatar.io](https://unavatar.io): Added [Pinterest](https://unavatar.io/pinterest), [Snapchat](https://unavatar.io/snapchat), [Discord](https://unavatar.io/discord), [Behance](https://unavatar.io/behance), [Threads](https://unavatar.io/threads), [Steam](https://unavatar.io/steam), [Tumblr](https://unavatar.io/tumblr), [Facebook](https://unavatar.io/facebook), [Flickr](https://unavatar.io/flickr), [Stack Overflow](https://unavatar.io/stackoverflow), [Buy Me a Coffee](https://unavatar.io/buymeacoffee), [Google Play](https://unavatar.io/google-play), [Apple Store](https://unavatar.io/apple-store), [Xbox Gamertag](https://unavatar.io/xboxgamertag), and [PSN Profiles](https://unavatar.io/psnprofiles) support.
 - [unavatar.io](https://unavatar.io): [GitHub](https://unavatar.io/github) now supports email avatar lookup.
 - [unavatar.io](https://unavatar.io): [Gravatar](https://unavatar.io/gravatar) now accepts pre-computed MD5/SHA256 hash as input.
@@ -382,7 +382,7 @@
 ### April 2021
 
 - [Microlink API](/docs/api/getting-started/overview): Added better default fonts and improved [font rendering](https://pro.microlink.io/?url=https%3A%2F%2Ffont-rendering.vercel.app&screenshot=&fullPage=&embed=screenshot.url).
-- [Microlink CLI](https://github.com/microlinkhq/cli): Released v2.0.0.
+- [Microlink CLI](https://github.com/microlinkhq/microlink): Released v2.0.0.
 - [Microlink Function](https://github.com/microlinkhq/function): Released [@microlink/local](https://github.com/microlinkhq/local) for running them locally.
 
 ### March 2021